added redux

+++ src/components/profile/Profile.js
@@ -1,28 +1,46 @@
 import React, { useState } from "react";
 import "./profile.css";
 
+import { useDispatch, useSelector } from "react-redux";
+import { setName } from "../../actions";
+
 import { Container, Row, Col, Form } from "react-bootstrap";
 import Image from "react-bootstrap/Image";
 
-import { userUpdate, deleteAccount } from "../../apiRoutes/UserRoutes";
+import {
+  disconnectAll,
+  userUpdate,
+  deleteAccount,
+} from "../../apiRoutes/UserRoutes";
+import CookieManager from "../../apiRoutes/cookieManager";
 import Buttons from "../subcomponents/Buttons";
 
 import profileImage from "../images/profil-picture-anonymous.png";
 
 function Profile(props) {
+  const dispatch = useDispatch();
+
+  const name = useSelector((state) => state.user.name);
   const [nameChange, setNameChange] = useState("");
   const [editStatus, setEditStatus] = useState(false);
 
   const onClickConfirm = () => {
-    if (nameChange && props.name !== nameChange) {
-      props.onNewName(nameChange);
+    if (nameChange && name !== nameChange) {
+      dispatch(setName(nameChange));
       userUpdate(nameChange);
     }
   };
 
-  const handleDeleteAccount = () => {
-    deleteAccount();
-    props.onDeleteAccount();
+  const disconnect = (deletion) => {
+    if (deletion === "deleteAccount") {
+      deleteAccount();
+    } else if (deletion === "deleteCookies") {
+      disconnectAll();
+    }
+
+    CookieManager("delete");
+    dispatch(setName(undefined));
+    props.onDisconnect();
   };
 
   return (
@@ -40,12 +58,12 @@ function Profile(props) {
           <Row>
             <Col xs={12} md={{ span: "4", offset: "2" }}>
               {!editStatus ? (
-                <h2 className="text-white text-center">{props.name}</h2>
+                <h2 className="text-white text-center">{name}</h2>
               ) : (
                 <Form.Control
                   className="bg-primary border-secondary text-light"
                   id="inputTask"
-                  placeholder={props.name}
+                  placeholder={name}
                   onChange={(e) => setNameChange(e.target.value)}
                 />
               )}
@@ -83,7 +101,7 @@ function Profile(props) {
             bgColor={"#1f6153"}
             textColor={"text-success"}
             description={"Disconnect"}
-            handleClick={() => props.onDisconnect()}
+            handleClick={disconnect}
           />
         </Col>
         <Col className="d-flex justify-content-center" xs={12}>
@@ -91,7 +109,7 @@ function Profile(props) {
             bgColor={"#4a3f77"}
             textColor={"text-info"}
             description={"Disconnect all devices"}
-            handleClick={() => props.onDisconnectAll()}
+            handleClick={() => disconnect("deleteCookies")}
           />
         </Col>
       </Row>
@@ -100,7 +118,7 @@ function Profile(props) {
           bgColor={"#773f3f"}
           textColor={"text-warning"}
           description={"Delete account"}
-          handleClick={handleDeleteAccount}
+          handleClick={() => disconnect("deleteAccount")}
         />
       </Row>
     </Container>
